Musicology.Grams

Documentation

(.:) :: (b -> c) -> (a1 -> a2 -> b) -> a1 -> a2 -> c

(.::) :: (b -> c) -> (a1 -> a2 -> a3 -> a4 -> b) -> a1 -> a2 -> a3 -> a4 -> c

gramsMach :: Int -> Process a [a]

grams :: Foldable t => Int -> t a -> [[a]]

skipgramsRMach :: (Num k, Ord k, Integral n, Monad m) => m Bool -> ([a] -> Bool) -> (a -> a -> k) -> n -> k -> ProcessT m a [a]

skipgramsRSMach :: (Num k, Ord k, Integral n, Monad m) => m Bool -> ([a] -> Bool) -> (a -> a -> k) -> n -> k -> ProcessT m a [a]

skipgramsMach :: (Num k, Ord k, Integral n) => ([a] -> Bool) -> (a -> a -> k) -> n -> k -> Process a [a]

skipgramsSMach :: (Num k, Ord k, Integral n) => ([a] -> Bool) -> (a -> a -> k) -> n -> k -> Process a [a]

skipgrams :: (Foldable t, Num k, Ord k, Integral n) => t a -> ([a] -> Bool) -> (a -> a -> k) -> n -> k -> [[a]]

skipgramsS :: (Foldable t, Num k, Ord k, Integral n) => t a -> ([a] -> Bool) -> (a -> a -> k) -> n -> k -> [[a]]

mkCoin :: (PrimMonad m, Integral n) => Double -> n -> Gen (PrimState m) -> m Bool

skipgramsR :: (Foldable t, Num k, Ord k, Integral n, PrimMonad m) => t a -> Double -> Gen (PrimState m) -> ([a] -> Bool) -> (a -> a -> k) -> n -> k -> m [[a]]

skipgramsRS :: (Foldable t, Num k, Ord k, Integral n, PrimMonad m) => t a -> Double -> Gen (PrimState m) -> ([a] -> Bool) -> (a -> a -> k) -> n -> k -> m [[a]]

enum :: (Num b, Enum b) => [a] -> [(a, b)]

indexCost :: Num a1 => (a2, a1) -> (a3, a1) -> a1

mune :: [[(b1, b2)]] -> [[b1]]

indexSkipgrams :: [a] -> Int -> Int -> [[a]]

indexSkipgramsR :: PrimMonad m => [a] -> Double -> Gen (PrimState m) -> Int -> Int -> m [[a]]

indexSkipgramsS :: [a] -> Int -> Int -> [[a]]

indexSkipgramsRS :: PrimMonad m => [a] -> Double -> Gen (PrimState m) -> Int -> Int -> m [[a]]